+++ resolver/src/lib.rs
@@ -111,7 +111,7 @@
 //! # }
 //! ```
 //!
-//! Generaally after a lookup in an asynchornous context, there would probably be a connection made to a server, for example:
+//! Generally after a lookup in an asynchornous context, there would probably be a connection made to a server, for example:
 //!
 //! ```c
 //! let result = io_loop.run(lookup_future.and_then(|ips| {